fix implicit cast

+++ srcs/ft_strchr_region.c
@@ -6,7 +6,7 @@
 /*   By: mhoosen <mhoosen@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/06/04 12:58:46 by mhoosen           #+#    #+#             */
-/*   Updated: 2018/06/06 10:34:41 by mhoosen          ###   ########.fr       */
+/*   Updated: 2018/06/27 16:38:25 by mhoosen          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ ssize_t	ft_strchr_region(const char *s, int c, size_t start, size_t end)
 	while (i < end)
 	{
 		if (s[i] == (char)c)
-			return (i);
+			return ((ssize_t)i);
 		i++;
 	}
 	return (-1);
